Moved my test functions into an actual JUnit test class. This test class tests my functions on minusTime and addTime.
