+++ src/test/java/seedu/modtracker/ModuleListTest.java
@@ -81,24 +81,4 @@ class ModuleListTest {
         assertEquals(expected + System.lineSeparator(), outContent.toString());
     }
 
-    @Test
-    public void addTime_ModuleWithoutActualWorkload_ModuleWithWorkloadAdded() {
-
-        ByteArrayOutputStream outContent = new ByteArrayOutputStream();
-        System.setOut(new PrintStream(outContent));
-        modulesTest.addTime("addtime CS3030 4 4", true, storage);
-        String expected = "4 hours are added to CS3030" + System.lineSeparator();
-        assertEquals(expected + System.lineSeparator(), outContent.toString());
-    }
-
-    @Test
-    public void minusTime_ModuleWithActualWorkload_ModuleWithWorkloadSubtracted() {
-
-        ByteArrayOutputStream outContent = new ByteArrayOutputStream();
-        System.setOut(new PrintStream(outContent));
-        modulesTest.minusTime("minustime CS3030 2 4", true, storage);
-        String expected = "2 hours are removed from CS3030" + System.lineSeparator();
-        assertEquals(expected + System.lineSeparator(), outContent.toString());
-    }
-
 }
