+++ test/data/modlist.txt
@@ -1,3 +1,4 @@
+addMod CS3040
 addExp CS3030 4
 addMod CS5000
 addExp CS3030 4
@@ -5,23 +6,10 @@ addMod CS5000
 addExp CS3030 4
 addMod CS5000
 addExp CS3030 4
-
-addExp CS3030 4
-deleteExp CS3030
-addMod CS5000
-deleteMod CS5000
+addmod CS3030
 addtime CS3030 4 4
 minustime CS3030 2 4
-addExp CS3030 4
-deleteExp CS3030
-addMod CS5000
-deleteMod CS5000
-addtime CS3030 4 4
-minustime CS3030 2 4
-addExp CS3030 4
-deleteExp CS3030
-addMod CS5000
-deleteMod CS5000
+addmod CS3030
 addtime CS3030 4 4
 minustime CS3030 2 4
 addExp CS3030 4
